:bug: Fix access to onDidChangeActiveTextEditor on older atom versions
This function 9 months old but it seems we're still having customers
running older versions.

In the autocorrect sidebar the fix imply to use
onDidChangeActivePaneItem and check that the item is a text editor in
order to render the diffs for that editor.

In the autocorrect status, the fix is just to switch to
onDidChangeActivePaneItem as we also want to clear the autocorrect
status whe the user switches to a non-editor panel.