:bug: Fix autocorrect status subscriptions never disposed
It was leading to errors in the specs where clicking once on the status
item was triggering a lot of show/hide action making the tests flaky.
Depending on how much back and forth the sidebar could end up in the
right state without us noticing the issue or in the wrong state leading
then to failures.