:green_heart: Make plugin pass all tests again
For now we don't use the health error returned by the connector client,
instead we let the legacy code handle the response when we have one.

Other failing tests were caused by the old helpers not modifying the sub
modules objects. Fixing them was simply a matter of replacing the
helpers call with the ones now provided by kite-connector