:bug: Fix flickers on status icon when typing in a file
When changing the status we were switching to indexing mode before 
requesting the file's status. So, for a file that is ready we were 
adding/removing an attribute on the icon leading to that flicker. Now we 
only add/remove that attribute when we get the response from the status 
endpoint

Fixes  #434